apartment package

+++ back_end/src/main/java/com/university/booking_university_project/modules/apartment/repository/ApartmentRepository.java
@@ -3,5 +3,4 @@ package com.university.booking_university_project.modules.apartment.repository;
 import com.university.booking_university_project.jpa.entity.Apartment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ApartmentRepository extends JpaRepository<Apartment, Long> {
-}
+public interface ApartmentRepository extends JpaRepository<Apartment, Integer> {}
